requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
appdirs==1.4.4
appnope==0.1.2; sys_platform == 'darwin'
argcomplete==1.12.3
argh==0.26.2
asgiref==3.4.1; python_full_version >= '3.6.0'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
backcall==0.2.0
beautifulsoup4==4.9.3
biocommons.seqrepo==0.6.4
bioutils==0.5.5; python_full_version >= '3.6.0'
boto3==1.18.32
botocore==1.21.32; python_full_version >= '3.6.0'
bs4==0.0.1
canonicaljson==1.4.0; python_version ~= '3.5'
certifi==2021.5.30
charset-normalizer==2.0.4; python_version >= '3'
click==8.0.1
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
configparser==5.0.2; python_full_version >= '3.6.0'
cssselect==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
decorator==5.0.9; python_version >= '3.5'
fake-useragent==0.1.11
fastapi==0.68.1
frozendict==2.0.6; python_full_version >= '3.6.0'
ga4gh.vrs[extras]==0.7.0rc3
ga4gh.vrsatile.pydantic==0.0.1
gffutils==0.10.1
h11==0.12.0; python_full_version >= '3.6.0'
hgvs==1.5.1
httptools==0.3.0
humanfriendly==9.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.2; python_version >= '3'
importlib-metadata==4.8.1; python_full_version >= '3.6.0'
inflection==0.5.1; python_version >= '3.5'
ipython==7.27.0; python_version >= '3.7'
jedi==0.18.0; python_full_version >= '3.6.0'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsonschema==3.2.0
lxml==4.6.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markdown==3.3.4; python_full_version >= '3.6.0'
matplotlib-inline==0.1.2; python_version >= '3.5'
numpy==1.21.2; python_version < '3.11' and python_version >= '3.7'
parse==1.19.0
parsley==1.3
parso==0.8.2; python_full_version >= '3.6.0'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.20; python_full_version >= '3.6.2'
psycopg2-binary==2.9.1; python_full_version >= '3.6.0'
ptyprocess==0.7.0
pydantic==1.8.2
pyee==8.2.2
pyfaidx==0.6.2
pygments==2.10.0; python_version >= '3.5'
pyppeteer==0.2.6; python_version < '4' and python_full_version >= '3.6.1'
pyquery==1.4.3
pyrsistent==0.18.0; python_full_version >= '3.6.0'
pysam==0.16.0.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-jsonschema-objects==0.3.10
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests-html==0.10.0; python_full_version >= '3.6.0'
requests==2.26.0
s3transfer==0.5.0; python_full_version >= '3.6.0'
seqrepo==0.0.0
simplejson==3.17.5; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.2.1; python_version >= '3'
sqlparse==0.4.1; python_version >= '3.5'
starlette==0.14.2; python_full_version >= '3.6.0'
tabulate==0.8.9
tqdm==4.62.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
traitlets==5.1.0; python_version >= '3.7'
typing-extensions==3.10.0.2
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.15.0
uvloop==0.16.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==9.1
yoyo-migrations==7.3.2
zipp==3.5.0; python_full_version >= '3.6.0'
